start implementing rate limiting

+++ src/main/java/com/image/quickimage/image/config/ImageProperties.java
@@ -10,8 +10,8 @@ public record ImageProperties(
               int min,
               int originalWidth,
               int originalHeight,
-             int maxQuality,
-             int minQuality,
-             int defaultQuality
+             float maxQuality,
+             float minQuality,
+             float defaultQuality
 ) {
 }
